o  more fixes for int Ids.
git-svn-id: https://svn.apache.org/repos/asf/maven/continuum/branches/id-refactor@498032 13f79535-47bb-0310-9956-ffa450edef68

+++ continuum-core/src/main/java/org/apache/maven/continuum/core/action/AddAssignableRolesAction.java
@@ -55,7 +55,7 @@ public class AddAssignableRolesAction
     public void execute( Map context )
         throws ContinuumException, ContinuumStoreException
     {
-        int projectGroupId =  getProjectGroupId( context );
+        long projectGroupId =  getProjectGroupId( context );
 
         ProjectGroup projectGroup = store.getProjectGroupWithBuildDetails( projectGroupId );
