+++ continuum-core/src/main/java/org/apache/maven/continuum/core/action/UpdateBuildDefinitionFromProjectAction.java
@@ -42,7 +42,7 @@ public class UpdateBuildDefinitionFromProjectAction
         throws Exception
     {
         BuildDefinition buildDefinition = getBuildDefinition( map );
-        int projectId = getProjectId( map );
+        long projectId = getProjectId( map );
 
         Project project = store.getProjectWithAllDetails( projectId );
